# openapi.yaml
openapi: 3.0.3
info:
  title: Sacsol API
  version: 1.0.0
  description: Inventory management API for Sacsol
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
paths:
  /api/audit-logs/:
    get:
      operationId: audit_logs_list
      description: Superadmin-only read-only listing of audit events.
      summary: List audit logs
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      tags:
      - audit-logs
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAuditLogList'
          description: ''
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
  /api/users/:
    get:
      operationId: users_list
      description: |-
        Super Admin–only management of users.
        Also exposes:
          - GET  /api/users/me/                 : current user profile (any authenticated user)
          - POST /api/users/me/password/        : change own password (any authenticated user)
          - POST /api/users/{id}/set-password/  : set another user's password (superadmin only)
      summary: List users
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      tags:
      - Users
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserBaseList'
          description: ''
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      operationId: users_create
      description: |-
        Super Admin–only management of users.
        Also exposes:
          - GET  /api/users/me/                 : current user profile (any authenticated user)
          - POST /api/users/me/password/        : change own password (any authenticated user)
          - POST /api/users/{id}/set-password/  : set another user's password (superadmin only)
      summary: Create user
      tags:
      - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
        required: true
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBase'
          description: ''
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /api/users/{id}/:
    get:
      operationId: users_retrieve
      description: |-
        Super Admin–only management of users.
        Also exposes:
          - GET  /api/users/me/                 : current user profile (any authenticated user)
          - POST /api/users/me/password/        : change own password (any authenticated user)
          - POST /api/users/{id}/set-password/  : set another user's password (superadmin only)
      summary: Get user
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this user.
        required: true
      tags:
      - Users
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBase'
          description: ''
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      operationId: users_update
      description: |-
        Super Admin–only management of users.
        Also exposes:
          - GET  /api/users/me/                 : current user profile (any authenticated user)
          - POST /api/users/me/password/        : change own password (any authenticated user)
          - POST /api/users/{id}/set-password/  : set another user's password (superadmin only)
      summary: Replace user
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this user.
        required: true
      tags:
      - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBase'
          description: ''
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    patch:
      operationId: users_partial_update
      description: |-
        Super Admin–only management of users.
        Also exposes:
          - GET  /api/users/me/                 : current user profile (any authenticated user)
          - POST /api/users/me/password/        : change own password (any authenticated user)
          - POST /api/users/{id}/set-password/  : set another user's password (superadmin only)
      summary: Update user
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this user.
        required: true
      tags:
      - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedUserUpdateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedUserUpdateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedUserUpdateRequest'
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBase'
          description: ''
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      operationId: users_destroy
      description: |-
        Super Admin–only management of users.
        Also exposes:
          - GET  /api/users/me/                 : current user profile (any authenticated user)
          - POST /api/users/me/password/        : change own password (any authenticated user)
          - POST /api/users/{id}/set-password/  : set another user's password (superadmin only)
      summary: Delete user
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this user.
        required: true
      tags:
      - Users
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '204':
          description: No content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /api/users/{id}/set-password/:
    post:
      operationId: users_set_password
      description: Super Admin can set a new password for a user.
      summary: Set user password
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this user.
        required: true
      tags:
      - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPasswordRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SetPasswordRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SetPasswordRequest'
        required: true
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '200':
          description: Password updated successfully
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /api/users/me/:
    get:
      operationId: users_me
      description: Return the currently authenticated user.
      summary: Get current user
      tags:
      - Users
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
          description: ''
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /api/users/me/password/:
    post:
      operationId: users_change_own_password
      description: Authenticated user can change their own password by providing current
        and new passwords.
      summary: Change pwn password
      tags:
      - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeOwnPasswordRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ChangeOwnPasswordRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ChangeOwnPasswordRequest'
        required: true
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '200':
          description: Password updated successfully
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
components:
  schemas:
    AuditLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        entry_id:
          type: string
          format: uuid
          readOnly: true
        user_username:
          type: string
          readOnly: true
        action:
          type: string
          readOnly: true
        changes:
          readOnly: true
          nullable: true
          type: string
        timestamp:
          type: string
          format: date-time
          readOnly: true
    ChangeOwnPasswordRequest:
      type: object
      properties:
        current_password:
          type: string
          writeOnly: true
          minLength: 1
        new_password:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - current_password
      - new_password
...
    Me:
      type: object
      description: Read-only representation of the current user.
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        email:
          type: string
          format: email
          title: Email address
          maxLength: 254
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        role:
          type: string
          readOnly: true
        is_active:
          type: boolean
          title: Active
          description: Designates whether this user should be treated as active. Unselect
            this instead of deleting accounts.
        date_joined:
          type: string
          format: date-time
          readOnly: true
      required:
      - username
    PaginatedAuditLogList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/AuditLog'
...
    PaginatedUserBaseList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/UserBase'
...
    PatchedUserUpdateRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          title: Email address
          maxLength: 254
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        is_active:
          type: boolean
          title: Active
          description: Designates whether this user should be treated as active. Unselect
            this instead of deleting accounts.
        role:
          enum:
          - staff
          - manager
          type: string
          description: |-
            * `staff` - staff
            * `manager` - manager
          x-spec-enum-id: 8f2a4513995fb1c3
    SetPasswordRequest:
      type: object
      properties:
        password:
          type: string
          writeOnly: true
          minLength: 6
      required:
      - password
...
    UserBase:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        email:
          type: string
          format: email
          title: Email address
          maxLength: 254
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        role:
          type: string
          readOnly: true
        is_active:
          type: boolean
          title: Active
          description: Designates whether this user should be treated as active. Unselect
            this instead of deleting accounts.
        date_joined:
          type: string
          format: date-time
          readOnly: true
      required:
      - username
    UserCreateRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        email:
          type: string
          format: email
          title: Email address
          maxLength: 254
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        role:
          enum:
          - staff
          - manager
          type: string
          description: |-
            * `staff` - staff
            * `manager` - manager
          x-spec-enum-id: 8f2a4513995fb1c3
          writeOnly: true
        password:
          type: string
          writeOnly: true
          minLength: 6
      required:
      - password
      - role
      - username
    UserUpdateRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          title: Email address
          maxLength: 254
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        is_active:
          type: boolean
          title: Active
          description: Designates whether this user should be treated as active. Unselect
            this instead of deleting accounts.
        role:
          enum:
          - staff
          - manager
          type: string
          description: |-
            * `staff` - staff
            * `manager` - manager
          x-spec-enum-id: 8f2a4513995fb1c3
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
- url: /